refactor:  배송 상태 변경 검증을 DeliveryValidator로 통합
